graph: Move `LogCode` to `util` mod and add GraphQL logs
Add GraphQL query logs
